Got the server running cleaned up the code and finshed up the mysql handler class
