+++ populator/src/mysqlHandler.java
@@ -64,6 +64,50 @@ Object[][] data = new Object [9][500];
 		return data;
 		
 		}
+	
+	public static void writeClassesToMysql(int id, String name, int g, String [] c){ //write the periods table to mysql
+		try{
+		String query = "insert into schedules (id, name, grade, pd1, pd2, pd3, pd4, pd5, pd6, pd7)"
+		        + " values (?, ?, ?, ?, ?, ?, ?, ?, ?, ?)"; 
+		java.sql.PreparedStatement preparedStmt = pop.conn.prepareStatement(query);	//sends statement to mysql
+		      preparedStmt.setInt (1, id);
+		      preparedStmt.setString(2, name);
+		      preparedStmt.setInt(3, g);
+		      preparedStmt.setString(4, c[0]);
+		      preparedStmt.setString(5, c[1]);
+		      preparedStmt.setString(6, c[2]);
+		      preparedStmt.setString(7, c[3]);
+		      preparedStmt.setString(8, c[4]);
+		      preparedStmt.setString(9, c[5]);
+		      preparedStmt.setString(10, c[6]);
+		      preparedStmt.execute();
+		
+	}catch(Exception e){ 
+		System.out.println("SOMETHING WENT WRONG");
+	}
+	}
+	
+	public static int addColumns(String period){
+		String query;
+		java.sql.PreparedStatement preparedStmt;
+		ResultSet rs1;
+
+		query = "SELECT  COUNT(" + period + ") FROM periods WHERE " + period + " = '1'"; //count amount of courses in a period where value is one (filled)
+		try {
+			preparedStmt = pop.conn.prepareStatement(query);
+			preparedStmt.execute(); //execute the statement
+			rs1 = preparedStmt.getResultSet(); //rs1 is the output from the query
+			rs1.absolute(1);  //gets result from the first row
+			int a= rs1.getInt(1); //gets the value of the first column's results added together
+			return a; //returns the total of the classes with "1" in that period
+		} catch (Exception e) { //catches an error
+			System.out.println(e);
+		}		
+	 return -1;	//returns -1 if there is an error
+	}
+	
+	
+	
 	}
